+++ src/containers/LoginContainer/LoginContainer.stories.tsx
@@ -1,4 +1,4 @@
-import { ComponentStory, ComponentMeta, ComponentStoryObj } from '@storybook/react';
+import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
 
 import LoginContainer from './LoginContainer';
 
@@ -7,14 +7,8 @@ export default {
   component: LoginContainer,
 } as ComponentMeta<typeof LoginContainer>;
 
-export const 소셜_로그인: ComponentStoryObj<typeof LoginContainer> = {
+export const 로그인: ComponentStoryObj<typeof LoginContainer> = {
   args: {
     step: 1,
-  },
-};
-
-export const 닉네임_입력: ComponentStoryObj<typeof LoginContainer> = {
-  args: {
-    step: 2,
   },
 };
