+++ src/containers/SignUpContainer/SignUpContainer.stories.tsx
@@ -1,0 +1,10 @@
+import { ComponentMeta, ComponentStoryObj } from '@storybook/react';
+
+import SignUpContainer from './SignUpContainer';
+
+export default {
+  title: 'Pages/회원가입',
+  component: SignUpContainer,
+} as ComponentMeta<typeof SignUpContainer>;
+
+export const 회원가입: ComponentStoryObj<typeof SignUpContainer> = {};
